# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "dssd-unlp-2021-grupo11-backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sociedadesAnonimas": {
            "get": {
                "tags": [
                    "sociedadAnonima-apoderado"
                ],
                "summary": "Sociedades anonimas",
                "description": "Sociedades anonimas del usuario logueado",
                "operationId": "getUserSociedadesAnonimas",
                "responses": {
                    "200": {
                        "description": "JSON con datos de la S.A.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/sociedadAnonima/{id}": {
            "get": {
                "tags": [
                    "sociedadAnonima-apoderado"
                ],
                "summary": "Sociedad anónima",
                "description": "Sociedad anonima con id",
                "operationId": "getUserSociedadAnonima",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON con datos de la S.A.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "sociedadAnonima-apoderado"
                ],
                "summary": "Corregir Sociedad Anonima rechazada por mesa de entradas",
                "description": "Corregir Sociedad Anonima rechazada por mesa de entradas",
                "operationId": "patchSociedadAnonima",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "S.A. actualizada",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "S.A. actualizada"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "No puedes corregir esta S.A.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "No puedes corregir esta S.A."
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/sociedadAnonimaByCaseId/{id}": {
            "get": {
                "tags": [
                    "sociedadAnonima-empleado"
                ],
                "summary": "Sociedad anonima",
                "description": "Sociedades anonima por caseId de Bonita",
                "operationId": "getSociedadAnonimaByCaseId",
                "parameters": [
                    {
                        "name": "bonitaCaseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON con datos de las S.A.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": ""
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "No tienes acceso a los datos de esta sociedad.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "No tienes acceso a los datos de esta sociedad."
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/updateSociedadAnonimaStatus/{taskId}": {
            "post": {
                "tags": [
                    "sociedadAnonima-empleado"
                ],
                "summary": "Sociedad anonima",
                "description": "Aprobar/Rechazar una tarea asignada, del empleado autenticado",
                "operationId": "updateSociedadAnonimaStatus",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "aprobado": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea aprobada/rechazada",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "Tarea aprobada/rechazada"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "No puedes aprobar/rechazar esta tarea",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "No puedes aprobar/rechazar esta tarea."
                            }
                        }
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Token Based",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}